@lanking520 2019-05-06
General thoughts, do you think it is nessary for us to have some real-time benchmarking on the performance once we do some upgrade like this?