@stu1130 2019-05-08
@lanking520 Sorry for late response. real-time benchmarking here means test performance in CI System? I would prefer running the bechmark across CUDA 9/9.2/10/10.1 on nightly build