@lanking520 2019-05-09
@perdasilva I saw a recent commit done by you to downgrade the CUDA version. Do you think this number is promising and we push it to master?